requirements.txt
# Core
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
httpx>=0.26.0

# Database
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
alembic>=1.13.1

# GCP & Integrations
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.16.0

# Document Generation (Phase 10)
openpyxl>=3.1.0
python-docx>=1.1.0
reportlab>=4.0.0

# Security (Phase 13+)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
